split Customer and Bank classes out to .cs files
