+++ Bank.cs
@@ -1,0 +1,8 @@
+namespace Millionaires
+{
+    public class Bank
+    {
+        public string Symbol { get; set; }
+        public string Name { get; set; }
+    }
+}
