+++ Customer.cs
@@ -1,0 +1,9 @@
+namespace Millionaires
+{
+    public class Customer
+    {
+        public string Name { get; set; }
+        public double Balance { get; set; }
+        public string Bank { get; set; }
+    }
+}
