+++ Program.cs
@@ -4,18 +4,6 @@ using System.Linq;
 
 namespace Millionaires
 {
-    public class Customer
-    {
-        public string Name { get; set; }
-        public double Balance { get; set; }
-        public string Bank { get; set; }
-    }
-
-    public class Bank
-    {
-        public string Symbol { get; set; }
-        public string Name { get; set; }
-    }
 
     class Program
     {
